updated dog classes and interfaces
