+++ webapp/src/dog/dogEnums.tsx
@@ -1,0 +1,54 @@
+import "date-fns";
+
+
+export enum BreedTypes {
+    "Labrador",
+    "German Shepherd",
+    "Bulldog",
+  }
+  export enum ColorTypes {
+    "Black",
+    "White",
+    "Brown",
+    "Ginger",
+    "Gold",
+    "Blue",
+    "Gray",
+    "Cream",
+    "Yellow",
+    "Other",
+  }
+  export enum HairTypes {
+    "Short",
+    "Medium",
+    "Long",
+  }
+  export enum SizeTypes {
+    "Small",
+    "Medium",
+    "Large",
+  }
+  export enum HeightTypes {
+    "Low",
+    "Medium",
+    "Tall",
+  }
+  export enum EarsTypes {
+    "Standing",
+    "Hanging",
+  }
+  export enum TailTypes {
+    "Long",
+    "Short",
+  }
+  export enum SpecialMarkTypes {
+    "Tattoo",
+    "Collar",
+    "Scar",
+    "None",
+  }
+  export enum BehavioursTypes {
+    "Shy",
+    "Energetic",
+    "Friendly",
+  }
